- merged the new location sharing code in HomeActivity.kt
