+++ app/build.gradle
@@ -35,6 +35,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.4.0'
     implementation 'com.google.android.material:material:1.4.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.2'
+    implementation 'com.google.firebase:firebase-messaging-ktx:23.0.0'
     testImplementation 'junit:junit:4.+'
     implementation 'com.google.android.material:material:1.5.0-beta01'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
