+++ app/src/main/java/com/pixieium/austtravels/utils/Constant.kt
@@ -11,4 +11,13 @@ object Constant {
     val MEGHNA_NOTIFICATION = "Meghna"
     val PADMA_NOTIFICATION = "Padma"
     val SHURMA_NOTIFICATION = "Shurma"
+
+    // home activity
+    val REQUEST_LIVE_TRACK = 0
+    val REQUEST_DIRECTIONS = 1
+    val REQUEST_SHARE_LOCATION = 2
+    val MSG_START_TIMER = 0
+    val MSG_STOP_TIMER = 1
+    val MSG_UPDATE_TIMER = 2
+    val NOTIFICATION_ID = 12044
 }
