+++ app/src/main/res/values/strings.xml
@@ -77,7 +77,6 @@
     <string name="return_back">RETURN BACK</string>
     <string name="forgot_pass_caption">We’ll send you a password reset link if the email you provide exists in our database</string>
     <string name="enter_your_phone">Enter your phone</string>
-    <string name="ping">PING</string>
 
     <!--location-->
     <string name="launch_activity">Launch activity</string>
@@ -87,9 +86,8 @@
     <string name="permission_denied_explanation">Permission was denied, but is needed for core functionality.</string>
     <string name="settings">Settings</string>
     <string name="ok">OK</string>
-    <string name="start_location_updates_button_text">Start receiving location updates</string>
-    <string name="stop_location_updates_button_text">Stop receiving location updates</string>
-    <string name="location_updated">Location Updated: %1$s</string>
+    <string name="stop_location_updates_button_text">Stop location sharing</string>
     <string name="preference_file_key">com.example.android.while_in_use_location.PREFERENCE_FILE_KEY</string>
     <string name="no_location_text">No current location</string>
+    <string name="sharing_location_prompt">You are currently sharing your location in the background</string>
 </resources>
